+++ probpipe/distributions/_sequential_joint.py
@@ -415,7 +415,7 @@ class SequentialJointDistribution(RecordDistribution, SupportsSampling, Supports
         result._conditioned_names = all_conditioned
         result._conditioned_values = {
             **self._conditioned_values,
-            **{k: jnp.asarray(v, dtype=jnp.float32) for k, v in observed.items()},
+            **{k: jnp.asarray(v) for k, v in observed.items()},
         }
         result._sampleable_error = self._compute_sampleable_error(
             result._conditioned_names, result._callable_parents,
